Refactoring site creation
This commit adds a build step to the article creation process. Akin
to a "lite" version of Jekyll or Gatsby, there is a template in
`assets/pages/_article.html` that is interpolated with content in
the `assets/articles` folder. In addition, header and footer
components now exist to create the necessary metadata for each article.