Fixing security issue with `imagemin` dependency
There was a security alert for `lodash.template`, which is a dependency of `imagemin`. This has
been updated in addition to all other dependencies. The corresponding code was updated to match
the new API for `imagemin` as well.

Also, there was a typo in the path for the hyperloop article that has been corrected. A build has
been run for the latest.